+++ modules/AWSPowerShell/Cmdlets/DirectConnect/Basic/Enable-DCTransitVirtualInterface-Cmdlet.cs
@@ -125,8 +125,8 @@ namespace Amazon.PowerShell.Cmdlets.DC
         #region Parameter NewTransitVirtualInterfaceAllocation_Mtu
         /// <summary>
         /// <para>
-        /// <para>The maximum transmission unit (MTU), in bytes. The supported values are 1500 and 9001.
-        /// The default value is 1500. </para>
+        /// <para>The maximum transmission unit (MTU), in bytes. The supported values are 1500 and 8500.
+        /// The default value is 1500 </para>
         /// </para>
         /// </summary>
         [System.Management.Automation.Parameter(ValueFromPipelineByPropertyName = true)]