+++ modules/AWSPowerShell/Cmdlets/Glue/Basic/Get-GLUECustomEntityTypeList-Cmdlet.cs
@@ -40,6 +40,17 @@ namespace Amazon.PowerShell.Cmdlets.GLUE
     public partial class GetGLUECustomEntityTypeListCmdlet : AmazonGlueClientCmdlet, IExecutor
     {
         
+        #region Parameter Tag
+        /// <summary>
+        /// <para>
+        /// <para>A list of key-value pair tags.</para>
+        /// </para>
+        /// </summary>
+        [System.Management.Automation.Parameter(ValueFromPipelineByPropertyName = true)]
+        [Alias("Tags")]
+        public System.Collections.Hashtable Tag { get; set; }
+        #endregion
+        
         #region Parameter MaxResult
         /// <summary>
         /// <para>
@@ -103,6 +114,14 @@ namespace Amazon.PowerShell.Cmdlets.GLUE
             }
             context.MaxResult = this.MaxResult;
             context.NextToken = this.NextToken;
+            if (this.Tag != null)
+            {
+                context.Tag = new Dictionary<System.String, System.String>(StringComparer.Ordinal);
+                foreach (var hashKey in this.Tag.Keys)
+                {
+                    context.Tag.Add((String)hashKey, (String)(this.Tag[hashKey]));
+                }
+            }
             
             // allow further manipulation of loaded context prior to processing
             PostExecutionContextLoad(context);
@@ -124,6 +143,10 @@ namespace Amazon.PowerShell.Cmdlets.GLUE
             if (cmdletContext.MaxResult != null)
             {
                 request.MaxResults = cmdletContext.MaxResult.Value;
+            }
+            if (cmdletContext.Tag != null)
+            {
+                request.Tags = cmdletContext.Tag;
             }
             
             // Initialize loop variant and commence piping
@@ -212,6 +235,7 @@ namespace Amazon.PowerShell.Cmdlets.GLUE
         {
             public System.Int32? MaxResult { get; set; }
             public System.String NextToken { get; set; }
+            public Dictionary<System.String, System.String> Tag { get; set; }
             public System.Func<Amazon.Glue.Model.ListCustomEntityTypesResponse, GetGLUECustomEntityTypeListCmdlet, object> Select { get; set; } =
                 (response, cmdlet) => response.CustomEntityTypes;
         }