+++ modules/AWSPowerShell/Cmdlets/Glue/Basic/New-GLUECustomEntityType-Cmdlet.cs
@@ -95,6 +95,17 @@ namespace Amazon.PowerShell.Cmdlets.GLUE
         public System.String RegexString { get; set; }
         #endregion
         
+        #region Parameter Tag
+        /// <summary>
+        /// <para>
+        /// <para>A list of tags applied to the custom entity type.</para>
+        /// </para>
+        /// </summary>
+        [System.Management.Automation.Parameter(ValueFromPipelineByPropertyName = true)]
+        [Alias("Tags")]
+        public System.Collections.Hashtable Tag { get; set; }
+        #endregion
+        
         #region Parameter Select
         /// <summary>
         /// Use the -Select parameter to control the cmdlet output. The default value is 'Name'.
@@ -175,6 +186,14 @@ namespace Amazon.PowerShell.Cmdlets.GLUE
                 WriteWarning("You are passing $null as a value for parameter RegexString which is marked as required. In case you believe this parameter was incorrectly marked as required, report this by opening an issue at https://github.com/aws/aws-tools-for-powershell/issues.");
             }
             #endif
+            if (this.Tag != null)
+            {
+                context.Tag = new Dictionary<System.String, System.String>(StringComparer.Ordinal);
+                foreach (var hashKey in this.Tag.Keys)
+                {
+                    context.Tag.Add((String)hashKey, (String)(this.Tag[hashKey]));
+                }
+            }
             
             // allow further manipulation of loaded context prior to processing
             PostExecutionContextLoad(context);
@@ -202,6 +221,10 @@ namespace Amazon.PowerShell.Cmdlets.GLUE
             if (cmdletContext.RegexString != null)
             {
                 request.RegexString = cmdletContext.RegexString;
+            }
+            if (cmdletContext.Tag != null)
+            {
+                request.Tags = cmdletContext.Tag;
             }
             
             CmdletOutput output;
@@ -267,6 +290,7 @@ namespace Amazon.PowerShell.Cmdlets.GLUE
             public List<System.String> ContextWord { get; set; }
             public System.String Name { get; set; }
             public System.String RegexString { get; set; }
+            public Dictionary<System.String, System.String> Tag { get; set; }
             public System.Func<Amazon.Glue.Model.CreateCustomEntityTypeResponse, NewGLUECustomEntityTypeCmdlet, object> Select { get; set; } =
                 (response, cmdlet) => response.Name;
         }