+++ modules/AWSPowerShell/Cmdlets/SecretsManager/Basic/Remove-SECSecret-Cmdlet.cs
@@ -87,7 +87,8 @@ namespace Amazon.PowerShell.Cmdlets.SEC
         /// don't use either, then by default Secrets Manager uses a 30 day recovery window.</para><para>Secrets Manager performs the actual deletion with an asynchronous background process,
         /// so there might be a short delay before the secret is permanently deleted. If you delete
         /// a secret and then immediately create a secret with the same name, use appropriate
-        /// back off and retry logic.</para><important><para>Use this parameter with caution. This parameter causes the operation to skip the normal
+        /// back off and retry logic.</para><para>If you forcibly delete an already deleted or nonexistent secret, the operation does
+        /// not return <code>ResourceNotFoundException</code>.</para><important><para>Use this parameter with caution. This parameter causes the operation to skip the normal
         /// recovery window before the permanent deletion that Secrets Manager would normally
         /// impose with the <code>RecoveryWindowInDays</code> parameter. If you delete a secret
         /// with the <code>ForceDeleteWithoutRecovery</code> parameter, then you have no opportunity