+++ modules/AWSPowerShell/Cmdlets/WAFV2/Basic/Get-WAF2RateBasedStatementManagedKey-Cmdlet.cs
@@ -28,13 +28,16 @@ using Amazon.WAFV2.Model;
 namespace Amazon.PowerShell.Cmdlets.WAF2
 {
     /// <summary>
-    /// Retrieves the keys that are currently blocked by a rate-based rule instance. The maximum
-    /// number of managed keys that can be blocked for a single rate-based rule instance is
-    /// 10,000. If more than 10,000 addresses exceed the rate limit, those with the highest
-    /// rates are blocked.
+    /// Retrieves the IP addresses that are currently blocked by a rate-based rule instance.
+    /// This is only available for rate-based rules that aggregate solely on the IP address
+    /// or on the forwarded IP address. 
     /// 
     ///  
     /// <para>
+    /// The maximum number of addresses that can be blocked for a single rate-based rule instance
+    /// is 10,000. If more than 10,000 addresses exceed the rate limit, those with the highest
+    /// rates are blocked.
+    /// </para><para>
     /// For a rate-based rule that you've defined inside a rule group, provide the name of
     /// the rule group reference statement in your request, in addition to the rate-based
     /// rule name and the web ACL name.